Load authorzation model for client
